created new run_sw file, and moved bulk of sw calls from interface file, as well as configure_code_run, to new file
